pyproject.toml
[project]
name = "all-work-of-openai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=3.4.3",
    "chainlit>=2.6.3",
    "openai>=1.98.0",
    "openai-agents>=0.2.3",
    "python-decouple>=3.8",
]
